Revert "feat: dockerfile improve, add build context feature (#320)"
This reverts commit c5a8322404a3dec9d853177875fab4c2659dda88.
