+++ backend/docs/docs.go
@@ -9579,36 +9579,6 @@ const docTemplate = `{
                 }
             }
         },
-        "github_com_raids-lab_crater_internal_util.VolumeMount": {
-            "type": "object",
-            "properties": {
-                "datasetID": {
-                    "type": "integer"
-                },
-                "mountPath": {
-                    "type": "string"
-                },
-                "subPath": {
-                    "type": "string"
-                },
-                "type": {
-                    "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeType"
-                }
-            }
-        },
-        "github_com_raids-lab_crater_internal_util.VolumeType": {
-            "type": "integer",
-            "enum": [
-                0,
-                1,
-                2
-            ],
-            "x-enum-varnames": [
-                "_",
-                "FileType",
-                "DataType"
-            ]
-        },
         "github_com_raids-lab_crater_pkg_crclient.ClusterNodeDetail": {
             "type": "object",
             "properties": {
@@ -10870,12 +10840,6 @@ const docTemplate = `{
                 },
                 "template": {
                     "type": "string"
-                },
-                "volumeMounts": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount"
-                    }
                 }
             }
         },
@@ -11189,7 +11153,7 @@ const docTemplate = `{
                 "volumeMounts": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount"
+                        "$ref": "#/definitions/internal_handler_vcjob.VolumeMount"
                     }
                 }
             }
@@ -11265,6 +11229,36 @@ const docTemplate = `{
                     "type": "string"
                 }
             }
+        },
+        "internal_handler_vcjob.VolumeMount": {
+            "type": "object",
+            "properties": {
+                "datasetID": {
+                    "type": "integer"
+                },
+                "mountPath": {
+                    "type": "string"
+                },
+                "subPath": {
+                    "type": "string"
+                },
+                "type": {
+                    "$ref": "#/definitions/internal_handler_vcjob.VolumeType"
+                }
+            }
+        },
+        "internal_handler_vcjob.VolumeType": {
+            "type": "integer",
+            "enum": [
+                0,
+                1,
+                2
+            ],
+            "x-enum-varnames": [
+                "_",
+                "FileType",
+                "DataType"
+            ]
         },
         "resource.Quantity": {
             "type": "object",
