+++ backend/docs/swagger.json
@@ -9571,36 +9571,6 @@
                 }
             }
         },
-        "github_com_raids-lab_crater_internal_util.VolumeMount": {
-            "type": "object",
-            "properties": {
-                "datasetID": {
-                    "type": "integer"
-                },
-                "mountPath": {
-                    "type": "string"
-                },
-                "subPath": {
-                    "type": "string"
-                },
-                "type": {
-                    "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeType"
-                }
-            }
-        },
-        "github_com_raids-lab_crater_internal_util.VolumeType": {
-            "type": "integer",
-            "enum": [
-                0,
-                1,
-                2
-            ],
-            "x-enum-varnames": [
-                "_",
-                "FileType",
-                "DataType"
-            ]
-        },
         "github_com_raids-lab_crater_pkg_crclient.ClusterNodeDetail": {
             "type": "object",
             "properties": {
@@ -10862,12 +10832,6 @@
                 },
                 "template": {
                     "type": "string"
-                },
-                "volumeMounts": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount"
-                    }
                 }
             }
         },
@@ -11181,7 +11145,7 @@
                 "volumeMounts": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount"
+                        "$ref": "#/definitions/internal_handler_vcjob.VolumeMount"
                     }
                 }
             }
@@ -11257,6 +11221,36 @@
                     "type": "string"
                 }
             }
+        },
+        "internal_handler_vcjob.VolumeMount": {
+            "type": "object",
+            "properties": {
+                "datasetID": {
+                    "type": "integer"
+                },
+                "mountPath": {
+                    "type": "string"
+                },
+                "subPath": {
+                    "type": "string"
+                },
+                "type": {
+                    "$ref": "#/definitions/internal_handler_vcjob.VolumeType"
+                }
+            }
+        },
+        "internal_handler_vcjob.VolumeType": {
+            "type": "integer",
+            "enum": [
+                0,
+                1,
+                2
+            ],
+            "x-enum-varnames": [
+                "_",
+                "FileType",
+                "DataType"
+            ]
         },
         "resource.Quantity": {
             "type": "object",
