+++ backend/internal/handler/image/interface.go
@@ -91,7 +91,7 @@ func NewImagePackMgr(conf *handler.RegisterConfig) handler.Manager {
 }
 
 var (
-	UserNameSpace   = config.GetConfig().Namespaces.Job
+	UserNameSpace   = config.GetConfig().Namespaces.Image
 	ProjectIsPublic = true
 	//nolint:mnd // default project quota: 20GB
 	GBit = int64(math.Pow(2, 30))
