+++ backend/pkg/packer/buildkit.go
@@ -12,19 +12,14 @@ import (
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/raids-lab/crater/internal/util"
 	"github.com/raids-lab/crater/pkg/config"
 )
 
 func (b *imagePacker) CreateFromDockerfile(c context.Context, data *BuildKitReq) error {
-	// Generate volumes and volumeMounts from request
-	volumes, volumeMounts, buildContext, err := b.generateVolumesAndMounts(c, data)
-	if err != nil {
-		return fmt.Errorf("failed to generate volumes and mounts: %w", err)
-	}
-
-	buildkitContainer := b.generateBuildKitContainer(data, volumeMounts, buildContext)
+	buildkitContainer := b.generateBuildKitContainer(data)
+	volumes := b.generateVolumes(data.JobName)
 	var configMap *corev1.ConfigMap
+	var err error
 	if configMap, err = b.createDockerfileConfigMap(c, data); err != nil {
 		return err
 	}
@@ -40,15 +35,7 @@ func (b *imagePacker) CreateFromDockerfile(c context.Context, data *BuildKitReq)
 	return nil
 }
 
-// generateVolumesAndMounts generates both volumes and volumeMounts based on the request
-// It reuses the resolveVolumeMount logic from vcjob package
-func (b *imagePacker) generateVolumesAndMounts(c context.Context, data *BuildKitReq) (
-	[]corev1.Volume,
-	[]corev1.VolumeMount,
-	map[string]string,
-	error,
-) {
-	// Initialize default volumes
+func (b *imagePacker) generateVolumes(jobName string) []corev1.Volume {
 	volumes := []corev1.Volume{
 		{
 			Name: "workspace",
@@ -75,59 +62,16 @@ func (b *imagePacker) generateVolumesAndMounts(c context.Context, data *BuildKit
 			VolumeSource: corev1.VolumeSource{
 				ConfigMap: &corev1.ConfigMapVolumeSource{
 					LocalObjectReference: corev1.LocalObjectReference{
-						Name: data.JobName,
+						Name: jobName,
 					},
 				},
 			},
 		},
 	}
-
-	// Initialize default volumeMounts for buildkit container
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      "harborcredits",
-			MountPath: "/.docker/config.json",
-			SubPath:   "config.json",
-			ReadOnly:  true,
-		},
-		{
-			Name:      "configmap-volume",
-			MountPath: "/workspace",
-			ReadOnly:  true,
-		},
-	}
-	buildContext := make(map[string]string)
-	// Process user-defined volume mounts
-	if len(data.VolumeMounts) > 0 {
-		pvcMap := make(map[string]bool) // Avoid duplicate PVC creation
-
-		for i, vm := range data.VolumeMounts {
-			// Reuse resolveVolumeMount logic from vcjob
-			volumeMount, err := util.ResolveVolumeMount(c, data.Token, vm, util.ImageCreate)
-			if err != nil {
-				return nil, nil, nil, fmt.Errorf("failed to resolve volume mount %s: %w", vm.MountPath, err)
-			}
-
-			// Add PVC volume if not already created
-			if !pvcMap[volumeMount.Name] {
-				volumes = append(volumes, util.CreateVolume(volumeMount.Name))
-				pvcMap[volumeMount.Name] = true
-			}
-
-			// Add volume mount
-			volumeMounts = append(volumeMounts, volumeMount)
-			buildContext[fmt.Sprintf("project%d", i)] = volumeMount.MountPath
-		}
-	}
-
-	return volumes, volumeMounts, buildContext, nil
+	return volumes
 }
 
-func (b *imagePacker) generateBuildKitContainer(
-	data *BuildKitReq,
-	volumeMounts []corev1.VolumeMount,
-	buildContext map[string]string,
-) []corev1.Container {
+func (b *imagePacker) generateBuildKitContainer(data *BuildKitReq) []corev1.Container {
 	output := fmt.Sprintf("type=image,name=%s,push=true", data.ImageLink)
 	archs := strings.Join(data.Archs, ",")
 
@@ -151,14 +95,8 @@ func (b *imagePacker) generateBuildKitContainer(
 		--node amd-node \
 		--driver remote tcp://%s:1234 && \
 		%s	docker buildx use multi-platform-builder && \
-		docker buildx build --progress plain --platform %s \
-		--file /workspace/Dockerfile --output %s \
-		--build-context project=/data/home/user/ \
+		docker buildx build --progress plain --platform %s --file /workspace/Dockerfile --output %s /workspace
 	`, buildkitdAmdNameSpace, appendArmCmd, archs, output)
-	for key, value := range buildContext {
-		cmd += fmt.Sprintf(" --build-context %s=%s", key, value)
-	}
-	cmd += " /workspace"
 
 	setupCommands := []string{
 		"/bin/sh",
@@ -177,7 +115,19 @@ func (b *imagePacker) generateBuildKitContainer(
 					Value: "/.docker",
 				},
 			},
-			VolumeMounts: volumeMounts,
+			VolumeMounts: []corev1.VolumeMount{
+				{
+					Name:      "harborcredits",
+					MountPath: "/.docker/config.json",
+					SubPath:   "config.json", // 只挂载文件，不覆盖目录
+					ReadOnly:  true,
+				},
+				{
+					Name:      "configmap-volume",
+					MountPath: "/workspace",
+					ReadOnly:  true,
+				},
+			},
 		},
 	}
 	return buildkitContainer
@@ -244,9 +194,6 @@ func (b *imagePacker) createJob(
 	jobMeta := metav1.ObjectMeta{
 		Name:      data.JobName,
 		Namespace: data.Namespace,
-		Labels: map[string]string{
-			"app": "image-create",
-		},
 		Annotations: map[string]string{
 			AnnotationKeyUserID:      fmt.Sprint(data.UserID),
 			AnnotationKeyImageLink:   data.ImageLink,
@@ -282,23 +229,6 @@ func (b *imagePacker) createJob(
 				EnableServiceLinks: ptr.To(false),
 				NodeSelector: map[string]string{
 					"kubernetes.io/arch": "amd64",
-				},
-				Affinity: &corev1.Affinity{
-					NodeAffinity: &corev1.NodeAffinity{
-						PreferredDuringSchedulingIgnoredDuringExecution: []corev1.PreferredSchedulingTerm{
-							{
-								Weight: 40,
-								Preference: corev1.NodeSelectorTerm{
-									MatchExpressions: []corev1.NodeSelectorRequirement{
-										{
-											Key:      "nvidia.com/gpu.present",
-											Operator: corev1.NodeSelectorOpDoesNotExist,
-										},
-									},
-								},
-							},
-						},
-					},
 				},
 			},
 		},
