+++ backend/pkg/packer/interface.go
@@ -7,7 +7,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/raids-lab/crater/dao/model"
-	"github.com/raids-lab/crater/internal/util"
 )
 
 type ImageRegistrySecret struct {
@@ -30,8 +29,6 @@ type BuildKitReq struct {
 	Template     string
 	BuildSource  model.BuildSource
 	Archs        []string
-	VolumeMounts []util.VolumeMount
-	Token        util.JWTMessage // Token information for volume resolution
 }
 
 type SnapshotReq struct {
