+++ frontend/src/services/api/imagepack.ts
@@ -212,11 +212,6 @@ export interface DockerfileCreate {
   template: string
   buildSource: ImagePackSource
   archs: string[]
-  volumeMounts?: {
-    type: number
-    subPath: string
-    mountPath: string
-  }[]
 }
 
 export interface EnvdCreate {
